feat(test): split plugin-cursor emit suite for SLOBAC M6
Add per-suite suiteTempDir helper and replace monolithic emit.test.ts with ten
domain Vitest modules matching the plugin-claude emit split pattern.

Mark M5 (plugin-claude discover split) checked in milestones.md to match the
completed sub-run. Memory bank refreshed for active M6: build done, QA/reflect pending.
