--- memory-bank/active/.qa-validation-status
@@ -1,9 +1,0 @@
-PASS
-task: slobac-audit-remediation-m5 (M5 split plugin-claude discover.tests)
-date: 2026-05-02
-phase: L2 QA
-
-Summary: Seven-domain split, discover-helpers parity with M4, monolith removed,
-discover It-count 58 and package-wide It-count 144 match plan; docs plugin tree
-lists discover-*.test.ts; no plugin-claude/src changes; split files free of TODO/
-debug noise.
