+++ memory-bank/active/activeContext.md
@@ -1,17 +1,15 @@
 # Active Context
 
-## Current Task: SLOBAC Audit Remediation — M5 (Split plugin-claude discover.test.ts)
+## Current Task: SLOBAC Audit Remediation — M6 (Split plugin-cursor emit.test.ts)
 
-**Phase:** L2 REFLECT — COMPLETE
+**Phase:** L2 BUILD — COMPLETE (awaiting QA)
 
 ## What Was Done
 
-- **Build:** Split completed. Added `packages/plugin-claude/test/test-support/discover-helpers.ts` (`discoverFixturesDir(import.meta.url)` → `test/fixtures`). New files: `discover-claude-md.test.ts`, `discover-simple-agent-skill.test.ts`, `discover-agent-ignore.test.ts`, `discover-manual-prompt.test.ts`, `discover-never-manual-prompt.test.ts`, `discover-agent-skill-io.test.ts`, `discover-rules.test.ts`. Removed `packages/plugin-claude/test/discover.test.ts`.
-- **Docs:** `packages/docs/docs/plugin-development/index.md` template tree updated to `discover-*.test.ts` / `emit-*.test.ts` (aligned with split suites).
-- **Verification:** Discover `it(` count **58**; package total **144**; full `pnpm test` green.
-- **QA (2026-05-02):** PASS — no corrective edits beyond memory-bank bookkeeping.
-- **Reflect (2026-05-02):** Insights recorded in `reflection-slobac-audit-remediation-m5.md`.
+- **Classify:** M6 classified Level 2; memory bank rewound from completed M5 per `/niko` Step 2a.
+- **Build:** Split `emit.test.ts` into ten domain files (`emit-global-prompt` through `emit-filename-case`) plus `test-support/emit-helpers.ts` (`suiteTempDir`).
+- **Verification:** Monorepo `pnpm test` green; plugin-cursor 137 tests (emit suite 62 `it` parity).
 
 ## Next Step
 
-L4 remediation project still has unchecked milestones (**M6** cursor emit, **M7** cursor discover). **`milestones.md` must advance only via `/niko`** (lifecycle rules). Invoke **`/niko`** when ready for the next milestone sub-run (`/niko-archive` only after full L4 capstone).
+Run **`/niko-qa`**, then **`/niko-reflect`**. On completion, **`/niko`** will check off M6 in `milestones.md` (lifecycle — do not hand-edit milestones).
